Use the same code to sort the success stories on the homepage section
